+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Zealot Barbarian 2024 Fixes.
 - Bigsby's Hand 2024 had stopped generating summons actors.
 - Wrath of the Sea range corrections.
+- 2024 Subclasses now generate spell list advancements.
 
 # 6.5.9